Move editor code to editor.hpp
